requirements.txt
fastapi
google-cloud-bigquery[async]
uvicorn
